Batch jobs (temp commit)
The job queue pool now supports batch jobs, which provide a range and
spawn one job per thread each with a slice of that range. The job
function can then use the provided range data to iterate through parts
of a collection and do something.

This is untested, and probably doesn't compile