Job counter (incomplete)
I've written a counter class that will store a job-push function, a job,
and an unsigned int and exposes an increment method that will add to an
atomic counter. If the counter reaches or exceeds the threshold (the
unsigned int) it will call the push function with the job as a
parameter.

Still to do: figure out how to handle increment amounts for batch jobs,
since they do more than one iteration per job, and figure out how to
handle updating the counter when its limit needs to change.