Starting, but not finishing, redesign
So I started some work, then realized that the whole design is sort of slightly fucked, so I went back to the drawing board and figured out a new, awesome design, but I don't want to implement it right now. So future me, look at your notebook next time you work on this, because it has all the designs.