Fixing JobCounters and Batching, breaking JobData
So counters are now fixed, they're a list of shared_ptrs on within Jobs. PushJob will now add a counter that deletes the Job's data, and PushBatchJob will add a counter to every Job in the batch that does the same, so the user now longer needs to worry about freeing JobData.

BatchJobData has also been rewritten as a wrapper for other JobData. This means that PushBatchJob can just create new wrappers and have them all reuse the original data pointer, preventing the slicing issue.

However, JobData has been broken; in my test case the PrintJobData that is supposed to be pointed to by all the batch jobs is falling out of scope at the end of QueuePrintJob, meaning not only can it not be used by the job function, but the counter that triggers after all the batches are done is now trying to delete already freed memory, which crashes the whole program. Closer, but we're still not there yet.

Also, the syntax for creating and pushingg Jobs is still disgusting, particularly when counters are involved. I need to write a helper function that hides the shared_ptrs rom the user, and further down the road I should think about a templated helper function that hides JobData creation in general, ideally allowing Jobs to be created with one function call for the job and the data and another for all the counters. That would be tops.