Reworking BatchJobData destruction
So I figured it out, and it was stupid. I had added a destructor to BatchJobData that would delete the wrapped data so that when the counter triggered it would only have to delete the BatchJobData. Naturally, this meant that when the BatchJobData I had created *on the stack* for QueuePrintJob fell out of scope the wrapped data was getting deleted with it because of that destructor. In the long run it's probably best to separate the pre- and post- split BatchJobData into their own classes, but for now I'm just going to remove that destructor and add a JobFunction that casts, deletes the wrapped data, then deletes the unwrapped data.

Everything now works great, and it seemingly runs without leaking.