Templatized Job creation
So jobs are now fully templatized and JobData is fully hidden from the end user, allowing jobs to be created simply by passing a function pointer and a set of parameters into a constructor. The implementation required significant changes to the overall structure of the system, so I'll outline the new process for creating a job and pushing it to the thread pool:

Job is now a template class that has template Ts..., which is a parameter pack. It derives from JobBase, which is empty besides a polymorphic destructor and an empty constructor, and is not templated. Job's constructor takes a pointer to a function that takes Ts..., and Ts.... In other words, it takes a set of arguments and a pointer to a function that also takes those arguments. It stores the pointer and a tuple containing the arguments.

The Job gets passed by reference to a templated overload of PushJob. This overload copies the Job to the heap and upcasts the resulting pointer to a JobBase pointer (type erasure). It then creates a GenericJob, which contains a JobBase pointer and a pointer to a function that takes a JobBase pointer (and a vector of JobCounters, but that system is about to receive a similar overhaul). The function pointer passed to the new GenericJob is of a templated function called RunJob that I'll explain in a minute. Once the GenericJob is created it gets passed to another overload of PushJob that actually pushes it into a JobQueue (this functionality is unchanged - GenericJob is just Job from previous revisions renamed).

When the GenericJob gets pulled off the queue, the RunJob function pointer is called and passed the upcast JobBase pointer. This is a templated function whose template argument is the same parameter pack that was used to create the Job. RunJob uses this parameter pack to downcast the JobBase pointer back to a Job pointer (type recovery) and uses an apply function (same as std::apply, but I can't use std::apply in my version of VS so I just copied the possible implementation) that forwards the tuple with the original arguments to the job function as actual arguments. When the GenericJob finishes, the JobCounter added by PushJob expires and triggers the deletion of the JobBase, and by extension the Job.

The end result of all of this is that the user can now write job functions exactly like normal functions and create Jobs with them with a single function call and no special syntax, and errors will happen at compile time rather than runtime. The downside is that batch jobs don't work yet and counters are still gross, but it's progress.