Templatized batch jobs pt1
So the batch job implementation is complete, and it would work exactly as I expected it to, except that the section and total section template parameters I wanted to pass to RunBatchJob are, obviously, not compile-time constants, and thus can't be used as template parameters. Duh. So I need to find some other way to pass those two numbers into RunBatchJob. My only thought so far is to use the current worker thread ID and the total number of worker threads, which should both be statically accessible from anywhere in the program, but that means RunBatchJob has to have logic for determining when there are fewer iterations than threads as well as PushJobAsBatch, AND it means that I can't shuffle batch jobs through the queues at all when that condition occurs. That is to say, if I want a batch of 3 but have 8 worker threads, I would have to always give the jobs to threads 0 through 2 rather than giving them to the current thread's three forward neighbors. Maybe that doesn't matter, but I feel like there's a better solution anyway, so I'm gonna wait till I find it.