Fixing Batching
So the answer to batching is to store the ID of the starting thread and
the number of sections in a subclass of Job<Ts...>. It gets stored when
the JobBase pointer is copy constructed, and RunJobAsBatch can then
downcast and pull out the starting thread ID and the number of sections
and use that to calculate the bounds for the current thread.

I'm kind of mad I didn't think of this sooner, it's so fucking simple.