Member function support pt1
So I've got an inheritance heirarchy now that hides invocation of a job behind a () operator. This should, in theory, allow free jobs and member jobs to be interchangeable, but I'm running into an issue with not being able to instantiate the abstract class they inherit from. I'm not entirely sure why this is occurring, as I'm only ever using pointers, not actual instances.

The good news is that member jobs get constructed and run just fine, and RunJob can be used to run either free or member jobs.