Duck-type revamp
So I've rewritten basically everything in a very similar but fundamentally different way; templates and duck-typing instead of OOP.

There are three main object categories now. The first two are Runnables and Callables. A Callable is any object that defines a () operator. function pointers, lambdas (capturing or not), etc., and I've even written a wrapper for member functions that can be used as a Callable. Runnables are Callables that define a () operator that takes no arguments.

Neither of those categories are actual classes; they are simply concepts that template arguments can adhere to.

The third category is Envelopes, which is an actual class. An Envelope is a type-erasing wrapper for a Callable; it contains a void pointer to the Callable and a pointer to a templated free function that takes a void pointer. The free function can cast to the provided template type and invoke the Callable. Envelopes define a () operator that calls the free function with the Callable, so they themselves are also Runnables.

Job<Ts...> has been rewritten as SimpleJob<Callable, Ts...>, and it now contains a Callable and the data tuple. A helper function has been added that can be used in lieu of the constructor to create SimpleJobs (the syntax is neater than using a constructor), and it will automatically use the MemberFunctionWrapper if a member function is given. JobQueue and JobQueuePool have been rewritten to work with SimpleJobs and Envelopes, and at this point they support standalone single-job queueing with any object that defines a () operator. Batching isn't working just yet.

JobCounters have also been rewritten as SealedEnvelopes. SealedEnvelope hides the shared_ptr to the dstructible object, which is the Seal. Seal contains an Envelope that is invoked in the Seal's destructor. Envelopes have a public method that can be used to add SealedEnvelopes as well, and self-deletion for standalone jobs is working as of this commit.

Overall I think this system is going to be far more flexible than an OOP approach, both or me and for the end-user.