Batching in the new system
So we now have SimpleJob and BatchJob. BatchJob is the same as SimpleJob, but it adds the starting thread ID and section count as well as a () operator that mimicks the functionality of the old RunJobAsBatch. Its constructor will also handle determining the number of sections, which PushJobAsBatch used to do.

The result is that all the necessary information for batching is encapsulated in a single object, and all that needs to happen after that is copying that object to the heap, creating an Envelope for each thread that points to it, and running those envelopes; everything else is handled in BatchJob's () operator.

This is a particularly nice solution because it will also allow for conversion between Simple- and BatchJobs via constructors later on.