Init function, major refactoring
So I've added an Init function that handles creating the worker threads and job queues and starts the system off on a job. It's templated such that the user just passes a callable object and any parameters, so the whole system is now good to go with one line of code.

I've also removed all the old classes, renamed all the files to what they should be named, and removed all unnecessary dependencies. Thanks to the heavy reliance on templates there's basically no dependency between most of the classes, and I've managed to remove all the ugly forward declarations and sketchy include logic; the whole thing is actually really clean now, and it builds super quick.

I still have to add some sugar for SealedEnvelopes, do namespacing to get the public interface solid, figure out how to build it such that it can actually be included and used in other projects, and add the fiberization, which is kind of the main selling point.

But this is pretty dang close to functional now.