Moving envelope functions into Envelope, Envelope constructor, console app
I've moved RunRunnable and DeleteRunnable into Envelope as static functions to reduce some of the clutter in globals. I've also added a templated constructor to Envelope that takes a pointer to an arbitrary type and constructs the envelope using the appropriate RunRunnable specialization for that type. This is just syntactic sugar to allow Envelopes to be used more intuitively used as a type-erasure mechanism for callable objects.

I've also changed the app to be a console app rather than a windows app. This allowed me to remove the console creation method and a lot of the cruft in main.