Fiberization & job calling
The primary feature is here: the user can now call jobs as parallel coroutines, freezing the current call stack and resuming it when the jobs finish.

The implementation is concise, but deceptively complex. CallJobs is called with a list of runnables. It breaks these into a vector of Envelopes using PackRunnablesNoAlloc, which is the same as PackRunnables but without any dynamic memory allocation (everything can stay on the stack since we're preserving the calling fiber), and passes them to an overload of CallJobs. This overload creates a job using FinishCalledJob and the address of the current fiber, puts it in a SealedEnvelope, and adds the SealedEnvelope to all the Envelopes in the vector. It then saves a static thread_local pointer to the vector on JobQueuePool (m_currentJobs), and then grabs a fiber to use and switches to it.

This is where the logic gets tricky. JobQueuePool has a static thread_local vector containing pointers to fibers for re-use. If there isn't a fiber in that vector, a new one will be created that starts at QueueJobsAndEnterJobLoop. When that fiber is switched to it will push the contents of obQueuePool::m_currentJobs, clear it, then enter the job loop.

Once all the jobs are finished the SealedEnvelope will expire, calling FinishCalledJob with the pointer to the original fiber, which was switched out of. FinishCalledJob will then add the current fiber's address to the vector of available fibers (m_availableFibers) and switch back to the old fiber.

However, FinishCalledJob does not end after the fiber switch. If there *is* a fiber in m_availableFibers during CallJob it will pop an address off it and switch into *that* fiber, which will then resume execution directly after the SwitchToiber call in FinishCalledJob. From here FinishCalledJob will push the contents of m_currentJobs, clear it, and then end, going back to the JobLoop.

So this is pretty spiffy. All job data lives on the stack. Every fiber gets reused, so after a while there will never be a need to allocate a new fiber. And best of all, there's no locking of any kind, because every interaction with the fiber system happens in a thread local way.