Reorganizing headers, parallel for
I've moved some includes from headers to cpps for classes that don't need the full implementations in their headers. This has helped compile times a bit, but since most of the substance is in templates I don't think it'll make much difference in the long run.

I've also implemented parallel for functionality. Nothing much to it, it works exactly as I expected and it took literally five minutes.

So Nova is just about feature complete now. All that's left is testing, poring over the code for small improvements or things I missed, and prepping it for use as an actual library.