BatchJob incrementing
BatchJobs will no longer pick a slice based on the ID of the thread they run on. Instead, a counter on the job will be atomically incremented each time a slice runs, and the value of the counter will determine the slice.

This is in preparation for work stealing; the old approach wouldn't have allowed a slice to be run on a different thread than it was originally added to.