Work stealing pt1
So each thread will now check every job queue while popping a job, and if it doesn't find a job in any of them it will sleep on a static condition variable in JobQueue rather than one specific to just one of the job queues. Pushing a job to any queue will wake one sleeping thread.

There's a problem with this approach. If every thread but one is working on a job and the remaining thread is spinning looking for a job, it will miss any jobs that are added to queues it has already checked, and it will sleep without grabbing any of those jobs. This means thosee jobs will not be grabbed until one of the other threads finishes their current job. This is exactly the same situation work stealing was meant to solve; jobs will potentially be blocked despite having threads available to process them.

Even more troubling, I think this can somehow lead to a race condition that completely halts the whole system. I've only seen it once, and haven't been able to replicate it, but... I think it's out there.

This whole piece of functionality is kind of gross anyway, I'm gonna rewrite it soon using some sort of atomic counter or interlocking to get threads to continue spinning so long as there's a job available in any queue.