Fixing access violation
So I'm still not entirely sure what was happening, but my best guess is that JobQueue::m_currentJobs was getting double deleted, once when FinishCalledJob cleared it and once when it fell off the stack after CallJobs.

This behavior was due to the fact that m_currentJobs was a pointer to a vector rather than a vector. I've rewritten it such that it is just a vector, and its contents will be swapped from the vector parameter of CallJobs with std::swap. This is essentially the same functionality I already had, save for the fact that the original vector parameter will no longer have ownership of the underlying data.

This has fixed the issue.