Removing MemberFunctionWrapper
So I discovered that the apply implementation I'm using can actually already handle member function pointers as long as the object to be used is the first element in the tuple. This meant SimpleJob was already equipped to handle member function pointers in exactly the syntax I was using, so I removed the MakeJob overload that handled inserting the MemberFunctionWrapper.

BatchJobs weren't so easy, as allowing them to maybe have an object reference as the first tuple element meant that their () operator would now maybe have to check elements 1 and 2 for the start and end rather than elements 0 and 1. But also maybe not. This was a problem given that the element indices have to be compile-time constants.

The solution was to look online for a type that would give me, as a compile-time constant, the first index of a given type within a tuple, which, amazingly, I was able to find. With that in hand all I had to do was write a couple member functions on BatchJob, one that returned a reference to the first unsigned int in a tuple and one that returned a reference to the second, and replace my std::get calls with calls to those member functions, giving me compile time access to the start and end values no matter where they were in the tuple.

From there I just tweaked BatchJob to store its parameters wholesale like SimpleJob and removed MemberFunctionWrapper completely.

There is one problem with this approach, however. Because the new member functions look for unsigned ints explicitly, the user can no longer pass in anything that isn't an unsigned int. Literals don't work, ints don't work, size_ts don't work, etc., unless they are explicitly wrapped in a cast to unsigned int. This bugged me until I realized that I could use it to ensure the user was marking the start and end parameters somehow. But just requiring them to cast to unsigned ints is too vague, so I wrote a typedef called BatchIndex that they can use instead which aliases unsigned int. Naturally, this doesn't prevent false positives on function signatures that happen to contain two or more adjacent unsigned ints, so I think that either aliasing a more obscure type or adding an assert that checks that the first unsigned int is either at index 0 or index 1 will be necessary, in addition to clearly documenting that the range should be the first two parameters of the job function.

But the upshot is that SimpleJob and BatchJob now have identical parameter types, making a cast easier to write, and that there's one less abstraction between the user's functions and the underlying INVOKE behavior. I'll call it a win.