The great race condition
Long story short, jobs were being finished before the call to PushEnvelopes in FinishCalledJob was finished, so the subsequent m_currentJobs.clear() was triggering the sealed envelope, which was sending the fiber back to the beginning of FinishCalledJob where it marked itself for reuse and switched back to the original fiber. This fiber switch happened in the middle of m_currentJobs' destructor, so it was never getting fully deallocated. This was probably why it was eventually causing bad deallocations; it was in some kind of invalid state.

I have a half solution, which is to swap m_currentJobs with a vector created in FinishCalledJobs and let the sealed envelopes be triggered by its destructor. The destructor still never finishes, which now causes a stack overflow since FinishCalledJob is essentially calling itself recursively through the sealed envelope, but at least nothing ever tries to use that stack allocated vector in its invalid state.

If I can figure out how to guarantee that I reach the job loop before the sealed envelope is triggered this will all be solved, I think.