Lots of nice improvements
Lots of stuff happening in this commit. First, BatchJob no longer has any special logic in its constructor; m_sections is initialized by the initializer list. Not a terribly significant change, but it's cleaner.

Second, I fixed another bug in the Call functionality; if Call is called with an empty Envelope vector the SealedEnvelope responsible for queueing FinishCalledJobs will do so immediately after the scope block it's created in ends because there are no jobs to keep it going. With multiple worker threads this is a problem because it means one of the other threads will immediately try to switch to the "original" fiber before it has been switched out of by the original thread. Really, Call being called with an empty Envelope vector should always be considered an error, so it will now cause an exception to be thrown. The only reason I found this bug at all is because of an error I made in implementing the BatchJob constructor change that caused m_sections to be wrongly initialized to 0. Lucky me.

Third, Push and Call now have overloads that take a vector of some runnable type, and Push has another overload that also takes a SealedEnvelope. Practically speaking these overloads probably won't be used much, but they make it easier for me to test, so I don't see any reason not to include them.

Fourth, Envelopes will now store only one SealedEnvelope rather than a vector of them. This is in response to my testing with the vector Push/Call overloads and discovering that copying the SealedEnvelopes into all the Envelopes takes a massive amount of time with large numbers of jobs, and it's resulted in a similarly massive speedup, increasing the throughput of those overloads by nearly 300% with empty jobs. At first glance it may seem like this prevents any single job from triggering multiple SealedEnvelopes (which honestly isn't a problem to begin with), but I could mimic that behavior by loading a SealedEnvelope with a lambda that calls other Envelopes.

The rest is really minor stuff. Where possible I've converted manual for loops to range-based for loops, I've added reserving for any place where I need a vector of a pre-determined size, push_back has been replaced by emplace_back where appropriate, and SealedEnvelopes are now passed only by reference until they get added to an Envelope to avoid as much overhead as possible reference counting.