Handling callbacks with Envelope rather than SealedEnvelope
Push will now take Envelopes as parameters and construct the SealedEnvelopes behind the scenes rather than taking SealedEnvelopes directly. To reflect this, the templated constructor has been removed from SealedEnvelope, and a constructor has been added to Envelope that will take a Runnable by value, copy it onto the heap, and set the run function to RunAndDeleteRunnable. This is an elegant way to allow the user to specify whether they should own the object or whether the Envelope should.