Removing Queue, cast improvements, namespace restructuring
The Queue class is no more; Push, PushBatch, and Call are now free functions in the Nova namespace, and all private methods and members are now free functions or static objects within Nova::internal. I've changed the PushEnvelope functions to overloads of Push and made them available outside of internal as well.

Simple- and BatchJob will now have public typedefs pointing to the other type with the correct template arguments. This allows for shorter, type-agnostic syntax for conversions between the two.

Things have been moved in and out of internal to reduce clutter in the Nova namespace. Simple- and BatchJob are now internal since there's no need to ever refer to them directly. SealedEnvelope is internal now as well, as Push's flow control handles them behind the scenes.