std::array implementation
This is a big one. PackRunnables has been heavily refactored, and now accepts both a std::array and a std::vector. It will use the vector to store envelopes created by batch jobs since each one spawns a variable number that can't be known at compile-time, but all other envelopes will be stored in the std::array to avoid dynamic memory allocation. The multi-job Push has been refactored to support this as well.

This means that, in theory and with compiler optimizations, a multi-job Push with a single job should be just as fast as a single job Push, so I've removed the single job Push overloads altogether.

I've also removed the build-Push because while it is convenient it necessitates additional overloads to differentiate between l- and rvalues.

So Push now has 5 overloads rather than 14.

Another small, but stupid, change is that the queue, the available fiber vector, and the currentJobs vector have been moved into a class called Resources. I was running into issues with Push(std::array<Envelope, N>) being in a different translation unit due to the template parameter and not being able to properly use the queue, and apparently static globals behave differently than static members. Hurray C++.