std::array for Call
Call will now store envelopes not generated by a batch job in a std::array rather than a vector. To accomplish this m_currentJobs has been replaced with m_callTrigger, which is a pointer to the SealedEnvelope that holds FinishCalledJob. The jobs get packed into the std::array and std::vector, the SealedEnvelope gets created and added to both, they're pushed, the SealedEnvelope is stripped from the Envelopes in the std::array and the std::vector gets cleared, then the fiber switches. After the switch m_callTrigger gets opened, leaving the Envelopes in the queue with the only remaining references to the SealedEnvelope.

This seems to have bumped the throughput up from ~500,000 jobs per second to ~3.6 million, so I'll call it a success.