Semaphore no more
So BlockingConcurrentQueue is really fucking slow. I'm not sure where moodycamel gets off saying it's only a little bit slower, it's reeeeeally fucking slow. We're talking four times as many jobs per second in the Push test and twice as many in the Call test by switching to regular old ConcurrentQueue, and it's all because of the semaphore. Semaphores just aren't lightweight, period, so tossing a semaphore onto a lockless data structure to prevent busy-waiting is a bit like removing the engine to prevent your car from overheating. No thank you.

Dummy locks and a condition variable with partial spinning is just as fast as busy-waiting in my tests, so I'm gonna go with that. It's required that I abstract the queue away from the internals of the job system and into its own wrapper class, which is good, since I was planning on doing that anyway.

So we're in a good place. Still need to figure out main-thread queueing though, and I'm still a long-ass way from being cross-platform, but we're in a good place.