Type trait for Envelope constructors
So with the help of stack overflow I figured out how to write a type trait to prevent Envelope's non-copy constructors from being called with an Envelope as a parameter. This was so that I could replace the value parameter with a reference parameter and get a nice speed boost, but it actually ended up being slower.

Bummer. But at least I've got some fancy looking, if completely unnecessary, type traits.