Queue adaptors
I've abstracted the queue out of QueueWrapper and into an adaptor class that contains the queue itself and defines Pop and Push methods to interact with it. The adaptor QueueWrapper uses is controlled by a define at the end of QueueAdaptors.h.

I tried to use this system to plug in a bounded, array-based MPMC queue I found on the internets (altered to move its elements on dequeue to allow SealedEnvelopes to trigger properly), and it worked, but the queue stops pulling jobs at some point and all the threads sleep. I'm hoping it's just the queue itself fucking up and not some kind of race condition, but what do I know.