Move-only Envelopes
I discovered an issue with Envelopes: if you copy then run a RunAndDelete Envelope the copy will be pointing to deleted memory.

A full redesign is on the docket for Envelopes, but in the meantime they've been made move-only. This isn't even a full solution, as the move constructor and assignment operator currently have no way to delete the currently held runnable. This is why a redesign is coming; I don't want to have six gazillion function pointers on each Envelope, there has to be a better way.