Envelope::m_deleteFunc
Envelope now has a function pointer for a templated delete function. It's been given a destructor that calls it, the move assignment operator will now call it before doing the move, and the standalone constructor will initialize it to DeleteRunnable.

All Envelope's members, including the delete pointer, now initialize properly (NoOp for the function pointers, nullptr for the runnable). Moves are now done with a special Move function to avoid code duplication. RunAndDeleteRunnable has been removed.