Dependent Push
Before each Envelope is run a thread_local pointer to its SealedEnvelope will be stored in Resources. I've written a new Push method, PushDependent, that will pass forward a boolean template parameter that causes the SealedEnvelope held by that pointer to be added to the Envelopes being queued. The implementation is about as simple as I expected, but I ran into some issues with overload resolution and constness that took some time to sort out.