Envelopes with shared ownership
RunRunnable now has an overload that accepts a std::shared_ptr as its Runnable. SplitBatchJob will now create a shared_ptr to the job and pass that to all the Envelopes rather than using an internalized SealedEnvelope to do the deletion.

This is necessary because with PushDependent in the picture that internal SealedEnvelope needs to always be available for a Call return.

The performance probably took a hit because of the additional dynamic allocation, but small object optimization will take care of that.