Envelope padding
Envelope is now padded based on the defined cache line size (NOVA_CACHE_LINE_BYTES). The moodycamel queue doesn't work with super-aligned types, so I'm padding it to 1.5x the defined size (rounded up) in order to guarantee no false-sharing. It's made a very slight difference in performance.

This required moving all of Envelope's data members into an inner class, which adds some syntactical gunk, but it's necessary for padding, so whatever.