Small object optimization
This was actually pretty easy.

The move constructor for job now checks whether the other job's runnable pointer is pointing into its padding array, and if it is it copies the padding array and sets the current job's runnable to point into its own padding array.

The padding array itself has been moved to the front of job to take advantage of job's alignment; as long as Runnable's alignment isn't more strict than job's we don't have to worry about aligning it at all.

I then added another constructor with an enable_if that ensures Runnable's alignment isn't more strict than job's and that its size isn't greater than the size of the padding. This constructor uses job::destruct_runnable, which calls the runnable's destructor but doesn't delete it, and placement news the runnable into the padding.

ezpz