moodycamel tokens pt1
This was a lot harder than I thought. Because the worker threads are created dynamically I couldn't just store two arrays or vectors of tokens, because adding workers would modify the collection, which isn't thread safe. And I couldn't store the tokens as thread_locals because the queue itself isn't static; the system needs to be able to support having multiple actual queues at once, which means having multiple sets of tokens per thread.

The solution is to give the tokens to the worker threads themselves.

Because queue-specific (everything moodycamel) and queue_wrapper-specific (main vs global, sleeping, etc.) code is supposed to be abstracted there had to be two sets of data: one that held the data needed for a single instance of the queue, and one that held the data needed for a single thread. These structs exist in moodycamel_adaptor and queue_wrapper, respectively. In init_thread a thread_data object is requested from the queue wrapper, which contains two queue_data objects that are requested from the moodycamel adaptor, and is stored on the worker thread object. Then a static thread_local queue_data pointer in queue_wrapper is set to point to that queue_data.

This way the queue wrapper, which is a singleton, can access the current thread's thread_data and pass the relevant queue_data to the adaptor's push and pop methods. The adaptor then passes the appropriate token from the queue_data to the queue's push and pop methods.

This has given us a solid ~80% speedup, and we're now looking at 20 million jobs per second in the push test and 33 million (!) in the call test, but there's a problem. When tokens are used with optimizations the memory usage explodes, topping out at 25mb by the end of the call test. It does this in both MSVC and clang, and it doesn't do it with optimizations disabled.

So the tokens are commented out until I can figure out why shit is borked.