Fixing potential issue with SMO size checking
So previously the enable_if that checked whether a Runnable type could be SMO'd checked the size of the Runnable type itself against the padding size, but because job_derived_smo<Runnable> is the type that's actually being written into the padding it should have been checking that to handle any implementation defined size discrepancies between the two, like (wait for it) THE VTABLE POINTER.

Herpaderpaderp

Unfortunately it looks like there's some issue with VS2015 that's preventing it from compiling nova with this change, so until I figure that out this is on a separate branch. Clang compiles it just fine.